backend/requirements.txt
# Backend dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0

# ML / RAG
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.21.0

# LLM
groq>=0.4.0

# Utilities
python-dotenv>=1.0.0
